8.3 obsluga wyjatku oraz poprawka zadania 7.6

+++ kodilla-stream/src/test/java/com/kodilla/stream/portfolio/BoardTestSuite.java
@@ -99,7 +99,7 @@ public class BoardTestSuite {
                 .reduce(0L, (sum, current) -> sum += current);
 
 
-        double average = taskSum/taskCount;
+        double average = taskSum/(double)taskCount;
 
 
         //Then
